requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.4
docutils==0.21.2
filelock==3.18.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.7
jnius==1.1.0
keyring==25.6.0
moviepy==1.0.3
Pillow==11.2.1
protobuf==6.31.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
redis==6.2.0
selenium==4.33.0
Sphinx==8.2.3
thread==2.0.5
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
yt_dlp==2025.6.9
